Organize tools into separate modules
