+++ segmentation/README.md
@@ -1,0 +1,10 @@
+# Segmentation Tool
+
+`drawSegment.py` offers a basic OpenCV interface for drawing polygonal ROIs on image slices.
+It can be used to collect interactive annotations for 3D volumes.
+
+## Running
+
+```bash
+python segmentation/drawSegment.py
+```
